dump

+++ srcs/nm/main.c
@@ -6,7 +6,7 @@
 /*   By: sclolus <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/20 17:02:35 by sclolus           #+#    #+#             */
-/*   Updated: 2017/12/06 17:17:18 by sclolus          ###   ########.fr       */
+/*   Updated: 2017/12/06 21:36:07 by sclolus          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,12 +29,12 @@ int	main(int argc, char **argv)
 			return (ft_error(2, (char*[]){"Failed to open(): ", nm_info->files[i]}, EXIT_FAILURE));
 		if (-1 == (fstat(fd, &file_stat)))
 			return (ft_error(2, (char*[]){"Failed to fstat(): ", nm_info->files[i]}, EXIT_FAILURE));
-		if (!(file_map = map_nm_file(fd, (uint64_t)file_stat.st_size)))
+		if (!(file_map = map_nm_file(fd, (size_t)file_stat.st_size)))
 			return (EXIT_FAILURE);
 		if (nm_info->nbr_files > 1)
 			printf("%s:\n", nm_info->files[i]);
-		nm(file_map, (uint64_t)file_stat.st_size, nm_info);
-		munmap(file_map, (uint64_t)file_stat.st_size);
+		nm(file_map, (size_t)file_stat.st_size, nm_info);
+		munmap(file_map, (size_t)file_stat.st_size);
 		close(fd);
 		i++;
 	}
